refactor(step2): menu 와 menugroup 패키지를 나누는 작업
메뉴라는 컨텍스트는 메뉴와 메뉴그룹이라는 도메인으로 나눌 수 있으므로
패키지도 이에 맞춰서 나눠 구현
