+++ src/main/java/kitchenpos/menus/application/dto/MenuCreateRequest.java
@@ -1,7 +1,5 @@
 package kitchenpos.menus.application.dto;
 
-import kitchenpos.menus.tobe.domain.Menu;
-
 import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
