+++ src/main/java/kitchenpos/menus/application/dto/MenuResponse.java
@@ -1,6 +1,6 @@
 package kitchenpos.menus.application.dto;
 
-import kitchenpos.menus.tobe.domain.Menu;
+import kitchenpos.menus.tobe.domain.menu.Menu;
 
 import java.math.BigDecimal;
 import java.util.List;
