+++ src/main/java/kitchenpos/menus/tobe/domain/menu/MenuDisplayedNameProfanities.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.tobe.domain.menu;
 
 public interface MenuDisplayedNameProfanities {
     boolean containsProfanity(String text);
