+++ src/main/java/kitchenpos/menus/tobe/domain/menu/MenuProduct.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.tobe.domain.menu;
 
 import javax.persistence.*;
 import java.math.BigDecimal;
