+++ src/main/java/kitchenpos/menus/tobe/domain/menu/MenuProductFactory.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.tobe.domain;
+package kitchenpos.menus.tobe.domain.menu;
 
 import kitchenpos.common.DomainService;
 import kitchenpos.products.tobe.domain.Product;
