+++ src/main/java/kitchenpos/menus/tobe/intrastructure/JpaMenuGroupRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.menus.tobe.intrastructure;
 
-import kitchenpos.menus.tobe.domain.MenuGroup;
-import kitchenpos.menus.tobe.domain.MenuGroupRepository;
+import kitchenpos.menus.tobe.domain.menugroup.MenuGroup;
+import kitchenpos.menus.tobe.domain.menugroup.MenuGroupRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
