+++ src/main/java/kitchenpos/menus/tobe/intrastructure/JpaMenuRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.menus.tobe.intrastructure;
 
-import kitchenpos.menus.tobe.domain.Menu;
-import kitchenpos.menus.tobe.domain.MenuRepository;
+import kitchenpos.menus.tobe.domain.menu.Menu;
+import kitchenpos.menus.tobe.domain.menu.MenuRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
