+++ src/main/java/kitchenpos/menus/ui/MenuGroupRestController.java
@@ -2,7 +2,7 @@ package kitchenpos.menus.ui;
 
 import kitchenpos.menus.application.MenuGroupService;
 import kitchenpos.menus.application.dto.MenuGroupCreateRequest;
-import kitchenpos.menus.tobe.domain.MenuGroup;
+import kitchenpos.menus.tobe.domain.menugroup.MenuGroup;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
