+++ src/test/java/kitchenpos/menus/application/MenuServiceTest.java
@@ -3,7 +3,8 @@ package kitchenpos.menus.application;
 import kitchenpos.menus.application.dto.MenuCreateRequest;
 import kitchenpos.menus.application.dto.MenuPriceChangeRequest;
 import kitchenpos.menus.application.dto.MenuProductCreateRequest;
-import kitchenpos.menus.tobe.domain.*;
+import kitchenpos.menus.tobe.domain.menu.*;
+import kitchenpos.menus.tobe.domain.menugroup.MenuGroupRepository;
 import kitchenpos.products.application.FakeProductEventPublisher;
 import kitchenpos.products.application.InMemoryProductRepository;
 import kitchenpos.products.application.ProductService;
