+++ src/test/java/kitchenpos/products/application/ProductServiceTest.java
@@ -1,8 +1,8 @@
 package kitchenpos.products.application;
 
 import kitchenpos.menus.application.InMemoryMenuRepository;
-import kitchenpos.menus.tobe.domain.Menu;
-import kitchenpos.menus.tobe.domain.MenuRepository;
+import kitchenpos.menus.tobe.domain.menu.Menu;
+import kitchenpos.menus.tobe.domain.menu.MenuRepository;
 import kitchenpos.products.tobe.domain.*;
 import kitchenpos.products.tobe.domain.ProductDisplayedNameProfanities;
 import kitchenpos.products.application.dto.ProductChangePriceRequest;
